Update: deleted id.py, and reorganise into attraction.py
